infer/requirements.txt
torch==1.11.0
torchvision==0.12.0
pytorch_lightning==1.6.3
numpy==1.21.6
Pillow==7.1.2
pandas==1.3.5